refactor: 테스트 파일 간 중복되는 CSV 파싱 로직 Util로 추출
